+++ api/app/Http/Controllers/Api/ProductImageController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\ProductResource;
 use App\Models\Image;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProductImageController extends Controller
 {
@@ -25,6 +26,8 @@ class ProductImageController extends Controller
 
     public function destroy(Product $product, Image $image)
     {
+        Gate::authorize('delete', $image);
+
         $product->images()->whereKey($image->id)->delete();
 
         return response()->noContent();
